+++ sdk/schemaregistry/schema-registry-avro/samples-dev/withEventHubsConsumerClient.ts
@@ -8,7 +8,11 @@
 import { DefaultAzureCredential } from "@azure/identity";
 import { SchemaRegistryClient, SchemaDescription } from "@azure/schema-registry";
 import { SchemaRegistryAvroEncoder } from "@azure/schema-registry-avro";
-import { EventHubConsumerClient, earliestEventPosition } from "@azure/event-hubs";
+import {
+  EventHubConsumerClient,
+  earliestEventPosition,
+  createEventDataAdapter,
+} from "@azure/event-hubs";
 
 // Load the .env file if it exists
 import * as dotenv from "dotenv";
@@ -70,7 +74,10 @@ export async function main() {
   await schemaRegistryClient.registerSchema(schemaDescription);
 
   // Create a new encoder backed by the client
-  const encoder = new SchemaRegistryAvroEncoder(schemaRegistryClient, { groupName });
+  const encoder = new SchemaRegistryAvroEncoder(schemaRegistryClient, {
+    groupName,
+    messageAdapter: createEventDataAdapter(),
+  });
 
   const eventHubConsumerClient = new EventHubConsumerClient(
     consumerGroup,