+++ sdk/schemaregistry/schema-registry-avro/samples-dev/withEventHubsProducerClient.ts
@@ -8,7 +8,7 @@
 import { DefaultAzureCredential } from "@azure/identity";
 import { SchemaRegistryClient, SchemaDescription } from "@azure/schema-registry";
 import { SchemaRegistryAvroEncoder } from "@azure/schema-registry-avro";
-import { EventHubProducerClient } from "@azure/event-hubs";
+import { EventHubProducerClient, createEventDataAdapter } from "@azure/event-hubs";
 
 // Load the .env file if it exists
 import * as dotenv from "dotenv";
@@ -75,6 +75,7 @@ export async function main() {
   // Create a new encoder backed by the client
   const encoder = new SchemaRegistryAvroEncoder(schemaRegistryClient, {
     groupName,
+    messageAdapter: createEventDataAdapter(),
   });
 
   const eventHubsProducerClient = new EventHubProducerClient(