+++ sdk/schemaregistry/schema-registry-avro/src/index.ts
@@ -1,10 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export {
-  SchemaRegistryAvroEncoder,
-  SchemaRegistryAvroEncoderOptions,
-} from "./schemaRegistryAvroEncoder";
+export { SchemaRegistryAvroEncoder } from "./schemaRegistryAvroEncoder";
 
 export * from "./models";
-export * from "./message";